baresip: Use UCI for basic service configuration
The baresip init script is changed to use UCI instead of the debian
style /etc/default basic service configuration.

A uci-defaults script is added to create a basic UCI config:
--
config baresip main
 option enable 0
 option options ''
--

With this change an issues is also resolved, on which the baresip
service doesn't start due to the baresip user not being the owner of
the /etc/baresip configuration directory.

Additionally a reload_config trigger is added for baresip.

Signed-off-by: Daniel Danzberger <dd@embedd.com>